Rename storage.rs to database.rs to reduce terminology variants
