+++ swap/src/bob/swap.rs
@@ -1,4 +1,4 @@
-use crate::{bob::event_loop::EventLoopHandle, state, storage::Database, SwapAmounts};
+use crate::{bob::event_loop::EventLoopHandle, database::Database, state, SwapAmounts};
 use anyhow::{bail, Result};
 use async_recursion::async_recursion;
 use rand::{CryptoRng, RngCore};
