+++ swap/src/lib.rs
@@ -20,10 +20,10 @@ pub mod alice;
 pub mod bitcoin;
 pub mod bob;
 pub mod cli;
+pub mod database;
 pub mod monero;
 pub mod network;
 pub mod state;
-pub mod storage;
 pub mod trace;
 
 pub type Never = std::convert::Infallible;
