+++ swap/src/main.rs
@@ -25,10 +25,10 @@ use swap::{
     bitcoin, bob,
     bob::swap::BobState,
     cli::{Command, Options, Resume},
+    database::Database,
     monero,
     network::transport::build,
     state::Swap,
-    storage::Database,
     trace::init_tracing,
     SwapAmounts,
 };
